Extract search result output partial
